+++ engine/durability.go
@@ -11,17 +11,17 @@ import "github.com/hupe1980/vecgo/metadata"
 // This interface intentionally matches the WAL method surface used by Tx.
 // A *wal.WAL satisfies this interface.
 type Durability interface {
-	LogPrepareInsert(id uint32, vector []float32, payload []byte, meta metadata.Metadata) error
-	LogCommitInsert(id uint32) error
+	LogPrepareInsert(id uint64, vector []float32, payload []byte, meta metadata.Metadata) error
+	LogCommitInsert(id uint64) error
 
-	LogPrepareBatchInsert(ids []uint32, vectors [][]float32, payloads [][]byte, metadataSlice []metadata.Metadata) error
-	LogCommitBatchInsert(ids []uint32) error
+	LogPrepareBatchInsert(ids []uint64, vectors [][]float32, payloads [][]byte, metadataSlice []metadata.Metadata) error
+	LogCommitBatchInsert(ids []uint64) error
 
-	LogPrepareUpdate(id uint32, vector []float32, payload []byte, meta metadata.Metadata) error
-	LogCommitUpdate(id uint32) error
+	LogPrepareUpdate(id uint64, vector []float32, payload []byte, meta metadata.Metadata) error
+	LogCommitUpdate(id uint64) error
 
-	LogPrepareDelete(id uint32) error
-	LogCommitDelete(id uint32) error
+	LogPrepareDelete(id uint64) error
+	LogCommitDelete(id uint64) error
 
 	// Close releases any resources held by the durability layer (e.g. file handles).
 	Close() error
@@ -36,18 +36,18 @@ type Durability interface {
 // correctness bifurcation.
 type NoopDurability struct{}
 
-func (NoopDurability) LogPrepareInsert(uint32, []float32, []byte, metadata.Metadata) error {
+func (NoopDurability) LogPrepareInsert(uint64, []float32, []byte, metadata.Metadata) error {
 	return nil
 }
-func (NoopDurability) LogCommitInsert(uint32) error { return nil }
-func (NoopDurability) LogPrepareBatchInsert([]uint32, [][]float32, [][]byte, []metadata.Metadata) error {
+func (NoopDurability) LogCommitInsert(uint64) error { return nil }
+func (NoopDurability) LogPrepareBatchInsert([]uint64, [][]float32, [][]byte, []metadata.Metadata) error {
 	return nil
 }
-func (NoopDurability) LogCommitBatchInsert([]uint32) error { return nil }
-func (NoopDurability) LogPrepareUpdate(uint32, []float32, []byte, metadata.Metadata) error {
+func (NoopDurability) LogCommitBatchInsert([]uint64) error { return nil }
+func (NoopDurability) LogPrepareUpdate(uint64, []float32, []byte, metadata.Metadata) error {
 	return nil
 }
-func (NoopDurability) LogCommitUpdate(uint32) error  { return nil }
-func (NoopDurability) LogPrepareDelete(uint32) error { return nil }
-func (NoopDurability) LogCommitDelete(uint32) error  { return nil }
+func (NoopDurability) LogCommitUpdate(uint64) error  { return nil }
+func (NoopDurability) LogPrepareDelete(uint64) error { return nil }
+func (NoopDurability) LogCommitDelete(uint64) error  { return nil }
 func (NoopDurability) Close() error                  { return nil }
